refactoring, creating header context
