+++ src/context/about.context.tsx
@@ -36,20 +36,15 @@ const presentationInitialValues : Presentation = {
 export interface AboutContextData {
     about : About ;
     profile: Presentation ;
-    menuActive : boolean;
-    isMenuActive: (isActive : boolean) => void;
 }
 export const AboutContext = createContext<AboutContextData>({
     about : {} as About,
-    profile : {} as Presentation,
-    menuActive : false,
-    isMenuActive: () => {}
+    profile : {} as Presentation
 })
 
 export const AboutProvider = () =>{
     const [about, setAbout] = useState<About>(aboutInitialValues);
     const [profile, setProfile] = useState<Presentation>(presentationInitialValues);
-    const [menuActive, setMenuActive] = useState<boolean>(false);
 
     useEffect( () => {
         const waitAbout = async() => {
@@ -62,16 +57,11 @@ export const AboutProvider = () =>{
 
     },[]);
 
-    const isMenuActive = (isActive : boolean) => {
-        // update experience active 
-        setMenuActive(!isActive);
-   }
 
     const value = {
         about,
         profile,
-        menuActive,
-        isMenuActive
+       
     }
 
     return (
